refactor: scheduler の boot 配線を scheduler-boot.ts に切り出す (#2015)
#2024 (PR #2025) と #2015 がそれぞれ server/index.ts に 1 行足した結果、
どちらも単独では超えないのに合わせると max-lines (600) を 1 行超えた。
main を取り込んだ時点で初めて赤くなる組み合わせで、CI が拾う前に手元で出た。

抑制はせず、この PR が触っている scheduler の boot 配線ごと切り出した。
各引数の理由のコメントが呼び出しより長く、index.ts の読み手が求めている
boot の「順序」ではないので、置き場所としても index.ts より適切。
index.ts は 601 → 584 行（計測値）になり余裕ができた。

boot 経路の変更なので実機で確認:
- 設定なし → systemTasks: 2、state は ~/.mulmoterminal/workspaces/<key>/ 配下
- 両方 off → systemTasks: 0、scheduler のファイルは作られない
- 両方 off + user task → userTasks: 1 で tick loop 起動
いずれも起動ディレクトリには何も作られない（#2024 の挙動も維持）。
